Added homepage polyfills.
